+++ iModelCore/ECDb/Tests/NonPublished/ProfileTests.cpp
@@ -202,7 +202,9 @@ TEST_F(ProfileTestFixture, CheckECDbProfileVersion)
             {ProfileVersion(4,0,0,1), ProfileState::Older(ProfileState::CanOpen::Readwrite, true)},
             {ProfileVersion(4,0,0,2), ProfileState::Older(ProfileState::CanOpen::Readwrite, true)},
             {ProfileVersion(4,0,0,3), ProfileState::Older(ProfileState::CanOpen::Readwrite, true)},
-            {ProfileVersion(4,0,0,4), ProfileState::UpToDate()},
+            {ProfileVersion(4,0,0,4), ProfileState::Older(ProfileState::CanOpen::Readwrite, true)},
+            {ProfileVersion(4,0,0,5), ProfileState::UpToDate()},
+            {ProfileVersion(4,0,0,6), ProfileState::Newer(ProfileState::CanOpen::Readwrite)},
             {ProfileVersion(4,0,1,0), ProfileState::Newer(ProfileState::CanOpen::Readwrite)},
             {ProfileVersion(4,0,1,1), ProfileState::Newer(ProfileState::CanOpen::Readwrite)},
             {ProfileVersion(4,0,1,99), ProfileState::Newer(ProfileState::CanOpen::Readwrite)},
@@ -220,9 +222,9 @@ TEST_F(ProfileTestFixture, CheckECDbProfileVersion)
         }
 
     std::vector<ProfileVersion> expectedTooOld = {ProfileVersion(3,6,99,0), ProfileVersion(3,7,0,0),ProfileVersion(3,7,0,1),ProfileVersion(3,7,3,1),ProfileVersion(3,7,3,2),ProfileVersion(3,7,4,3),ProfileVersion(3,100,0,0), ProfileVersion(3,100,0,1), ProfileVersion(3,100,1,1)};
-    std::vector<ProfileVersion> expectedOlderReadWriteAndUpgradable = {ProfileVersion(4,0,0,0), ProfileVersion(4,0,0,1), ProfileVersion(4,0,0,2), ProfileVersion(4,0,0,3)};
-    ProfileVersion expectedUpToDate = ProfileVersion(4,0,0,4);
-    std::vector<ProfileVersion> expectedNewerReadWrite = {ProfileVersion(4,0,0,5), ProfileVersion(4,0,1,0), ProfileVersion(4,0,1,3), ProfileVersion(4,0,2,0), ProfileVersion(4,0,10,99)};
+    std::vector<ProfileVersion> expectedOlderReadWriteAndUpgradable = {ProfileVersion(4,0,0,0), ProfileVersion(4,0,0,1), ProfileVersion(4,0,0,2), ProfileVersion(4,0,0,3), ProfileVersion(4,0,0,4)};
+    ProfileVersion expectedUpToDate = ProfileVersion(4,0,0,5);
+    std::vector<ProfileVersion> expectedNewerReadWrite = {ProfileVersion(4,0,0,6), ProfileVersion(4,0,1,0), ProfileVersion(4,0,1,3), ProfileVersion(4,0,2,0), ProfileVersion(4,0,10,99)};
     std::vector<ProfileVersion> expectedNewerReadonly = {ProfileVersion(4,1,1,0), ProfileVersion(4,1,0,0)};
     std::vector<ProfileVersion> expectedTooNew = {ProfileVersion(5,1,0,0), ProfileVersion(99,0,0,0)};
 
@@ -390,7 +392,7 @@ TEST_F(ProfileTestFixture, ImportRequiresVersionCustomAttribute)
     ASSERT_EQ(BentleyStatus::ERROR, ImportSchema(schema));
     auto lastIssue = issueListener.GetIssue();
     ASSERT_TRUE(lastIssue.has_value()) << "Should raise an issue.";
-    ASSERT_STREQ("ECSchema Schema1.01.00.01 requires ECDb version 999.9.9.9, but the current runtime version is only 4.0.0.4.", lastIssue.message.c_str());
+    ASSERT_STREQ("ECSchema Schema1.01.00.01 requires ECDb version 999.9.9.9, but the current runtime version is only 4.0.0.5.", lastIssue.message.c_str());
     }
 
     CloseECDb();
@@ -503,7 +505,7 @@ TEST_F(ProfileTestFixture, ApplyImportRequiresVersionToExistingSchema)
     ASSERT_EQ(BentleyStatus::ERROR, ImportSchema(schema));
     auto lastIssue = issueListener.GetIssue();
     ASSERT_TRUE(lastIssue.has_value()) << "Should raise an issue.";
-    ASSERT_STREQ("ECSchema Schema1.01.00.03 requires ECDb version 999.9.9.9, but the current runtime version is only 4.0.0.4.", lastIssue.message.c_str());
+    ASSERT_STREQ("ECSchema Schema1.01.00.03 requires ECDb version 999.9.9.9, but the current runtime version is only 4.0.0.5.", lastIssue.message.c_str());
     }
 
     CloseECDb();